third_party/blink/web_tests/platform/mac-mac10.13/external/wpt/service-workers/service-worker/navigation-timing-extended.https-expected.txt

This is a testharness.js-based test.
FAIL Service worker controlled navigation timing assert_greater_than: workerStart marking should not wait for worker activation to finish expected a number greater than 1634043126416.2002 but got 1634043126415.5
Harness: the test ran to completion.
